Revert "Implement ma"
