+++ App.js
@@ -1,23 +1,29 @@
-import * as React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import Home from "./components/Tabs/Home/home"
-import Friends from "./components/Tabs/Friends/friends"
-import Activity from "./components/Tabs/Activity/activity"
-import Profile from "./components/Tabs/Profile/profile"
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View, Image } from 'react-native';
 
-
-const Tab = createBottomTabNavigator();
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  tinyLogo: {
+    width: 50,
+    height: 50,
+  },
+  logo: {
+    width: 66,
+    height: 58,
+  },
+});
 
 export default function App() {
   return (
-    <NavigationContainer>
-      <Tab.Navigator>
-        <Tab.Screen name="Home" component={Home} />
-        <Tab.Screen name="Activity" component={Activity} />
-        <Tab.Screen name="Friends" component={Friends} />
-        <Tab.Screen name="Profile" component={Profile} />
-      </Tab.Navigator>
-    </NavigationContainer>
+    <View style={styles.container}>
+      <Image style={styles.logo} source={require('./assets/S.png')}/>
+      <Text>Welcome to Sitdown you fuckers.</Text>
+      <StatusBar style="auto" />
+    </View>
   );
 }
