--- components/Tabs/Activity/activity.js
@@ -1,29 +1,0 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View, Image } from 'react-native';
-
-
-const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      backgroundColor: '#fff',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    tinyLogo: {
-      width: 50,
-      height: 50,
-    },
-    logo: {
-      width: 66,
-      height: 58,
-    },
-  });
-  
-  export default function Activity() {
-    return (
-      <View style={styles.container}>
-        <Text>Activity</Text>
-        <StatusBar style="auto" />
-      </View>
-    );
-  }
