--- components/Tabs/Friends/friendList.js
@@ -1,30 +1,0 @@
-import { StyleSheet, View, Text} from 'react-native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-
-const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      backgroundColor: '#fff',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    tinyLogo: {
-      width: 50,
-      height: 50,
-    },
-    logo: {
-      width: 66,
-      height: 58,
-    },
-  });
-
-  const Stack = createNativeStackNavigator();
-
-  
-  export default function FriendList() {
-    return (
-        <View>
-            <Text>FriendList</Text>
-        </View>
-    );
-  }
