--- components/Tabs/Friends/friendProfile.js
@@ -1,31 +1,0 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View, Image } from 'react-native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-
-const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      backgroundColor: '#fff',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    tinyLogo: {
-      width: 50,
-      height: 50,
-    },
-    logo: {
-      width: 66,
-      height: 58,
-    },
-  });
-
-  const Stack = createNativeStackNavigator();
-
-  
-  export default function FriendProfile() {
-    return (
-        <View>
-            <Text>FriendProfile</Text>
-        </View>
-    );
-  }
